refactor: filterUrl 함수 분리
