+++ client/src/pages/Category/Category.js
@@ -12,6 +12,7 @@ import Card from '../../components/Card/Card';
 import Filter from './Filter';
 import HoverButton from '../../common/Buttons/HoverButton';
 // import Pagination from '../../components/Pagination/Pagination';
+import useFilterurl from '../../components/FIlterUrl/Filterurl';
 
 export default function Category() {
   //배포이후 baseUrl
@@ -28,7 +29,7 @@ export default function Category() {
     descendingOrder: true,
     sortType: sortTypeData[0].type,
   });
-  //console.log(filterCondtion);
+
   //현재 페이지 인덱스
   const [currentPage, setCurrentPage] = useState(0);
   const [cocktailData, setCocktailData] = useState([]);
@@ -39,35 +40,7 @@ export default function Category() {
 
   useEffect(() => {
     const fetchCocktails = async () => {
-      console.log('동작');
-      console.log(filterCondtion);
-      //클릭한 페이지
-      const page = currentPage + 1;
-      //카테고리 전체보기일때는 빈문자열
-      const categoryQuery =
-        filterCondtion.category === 'all'
-          ? ''
-          : `category=${filterCondtion.category}&`;
-      //맛 태그
-      const tasteTagQuery =
-        filterCondtion.tasteTag.length === 0
-          ? ''
-          : `,${filterCondtion.tasteTag.join(',')}`;
-
-      // 정렬 조건 내림차순 + 조회수= 조회수 높은 순
-      const getSortType = (descending, type) => {
-        if (descending) {
-          return type === 'viewed' ? 'most_viewed' : 'highest_rate';
-        }
-        return type === 'viewed' ? 'least_viewed' : 'lowest_rate';
-      };
-
-      const sort = getSortType(
-        filterCondtion.descendingOrder,
-        filterCondtion.sortType
-      );
-
-      const url = `${BASE_URL}cocktails/filter?${categoryQuery}tag=${filterCondtion.frequencyTag}${tasteTagQuery}&page=${page}&size=16&sort=${sort}`;
+      const url = useFilterurl(BASE_URL, currentPage, filterCondtion);
 
       try {
         const response = await fetch(url);
